separate all non CLI tests
